+++ lib/common/screens/scaffold.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+class ScaffoldScreen extends StatelessWidget {
+  const ScaffoldScreen({
+    required this.child,
+    this.appBar,
+    super.key,
+  });
+
+  final Widget child;
+  final PreferredSizeWidget? appBar;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: appBar,
+      body: Container(
+        decoration: BoxDecoration(
+          gradient: LinearGradient(
+            begin: Alignment.topCenter,
+            end: Alignment.bottomCenter,
+            colors: [
+              Colors.yellow.shade200,
+              Colors.pink.shade200,
+            ],
+          ),
+        ),
+        padding: const EdgeInsets.only(top: 10),
+        child: child,
+      ),
+    );
+  }
+}
