+++ lib/common/widgets/album_playlist.dart
@@ -2,6 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:on_audio_query/on_audio_query.dart';
 
 import '../models/music_settings_model.dart';
+import 'app_bar_widget.dart';
+import '../screens/scaffold.dart';
+import 'leading_icon.dart';
 
 class AlbumPlayList extends StatefulWidget {
   const AlbumPlayList({
@@ -34,67 +37,39 @@ class _AlbumPlayListState extends State<AlbumPlayList> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        decoration: BoxDecoration(
-          gradient: LinearGradient(
-            begin: Alignment.topCenter,
-            end: Alignment.bottomCenter,
-            colors: [
-              Colors.yellow.shade200,
-              Colors.pink.shade200,
-            ],
-          ),
-        ),
-        padding: const EdgeInsets.only(top: 40),
+    return ScaffoldScreen(
+      child: Padding(
+        padding: const EdgeInsets.all(8.0),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Row(
-                children: [
-                  IconButton(
-                    icon: const Icon(Icons.chevron_left_rounded),
-                    onPressed: () {
-                      Navigator.pop(context);
-                    },
-                  ),
-                  const SizedBox(
-                    width: 8,
-                  ), // Add some space between the icon and playlist name
-                  Text(
-                    widget.album.album,
-                    style: const TextStyle(
-                      fontSize: 20,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ],
+            AppBarWidget(
+              leadingIcon: const Center(
+                child: Icon(Icons.abc),
               ),
+              title: widget.album.album,
+              leadingPress: () {
+                Navigator.pop(context);
+              },
+              subtitle: '${widget.album.artist} | ${songs.length} song',
             ),
             Padding(
-              padding: const EdgeInsets.fromLTRB(
-                  25.0, 0, 8.0, 8.0), // Adjust the left padding here
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Text(
-                    'Songs (${songs.length})',
-                    style: const TextStyle(
-                      fontSize: 16,
-                    ),
-                  ),
-                ],
+              padding: const EdgeInsets.only(left: 20.0),
+              child: Text(
+                'Songs (${songs.length})',
+                style: Theme.of(context).textTheme.titleLarge,
               ),
             ),
             Expanded(
               child: ListView.builder(
-                padding: const EdgeInsets.fromLTRB(25.0, 0, 8.0, 8.0),
+                // padding: const EdgeInsets.fromLTRB(25.0, 0, 8.0, 8.0),
                 itemCount: songs.length,
                 itemBuilder: (context, index) {
                   final song = songs[index];
                   return ListTile(
+                    leading: const LeadingIcon(
+                      icon: Icon(Icons.music_note),
+                    ),
                     title: Text(song.title),
                     subtitle: Text(song.artist ?? ''),
                     onTap: () {
