+++ lib/common/widgets/artist_playlist.dart
@@ -1,7 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:on_audio_query/on_audio_query.dart';
+import 'package:sound_spin/common/screens/scaffold.dart';
 
 import '../models/music_settings_model.dart';
+import 'app_bar_widget.dart';
+import 'leading_icon.dart';
 
 class ArtistPlayList extends StatefulWidget {
   const ArtistPlayList({
@@ -34,67 +37,37 @@ class _ArtistPlayListState extends State<ArtistPlayList> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        decoration: BoxDecoration(
-          gradient: LinearGradient(
-            begin: Alignment.topCenter,
-            end: Alignment.bottomCenter,
-            colors: [
-              Colors.yellow.shade200,
-              Colors.pink.shade200,
-            ],
-          ),
-        ),
-        padding: const EdgeInsets.only(top: 40),
+    return ScaffoldScreen(
+      child: Padding(
+        padding: const EdgeInsets.all(8.0),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Row(
-                children: [
-                  IconButton(
-                    icon: const Icon(Icons.chevron_left_rounded),
-                    onPressed: () {
-                      Navigator.pop(context);
-                    },
-                  ),
-                  const SizedBox(
-                    width: 8,
-                  ), // Add some space between the icon and playlist name
-                  Text(
-                    widget.artist.artist,
-                    style: const TextStyle(
-                      fontSize: 20,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ],
+            AppBarWidget(
+              leadingIcon: const Center(
+                child: Icon(Icons.abc),
               ),
+              title: widget.artist.artist,
+              leadingPress: () {
+                Navigator.pop(context);
+              },
             ),
             Padding(
-              padding: const EdgeInsets.fromLTRB(
-                  25.0, 0, 8.0, 8.0), // Adjust the left padding here
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Text(
-                    'Songs (${songs.length})',
-                    style: const TextStyle(
-                      fontSize: 16,
-                    ),
-                  ),
-                ],
+              padding: const EdgeInsets.only(left: 20.0),
+              child: Text(
+                'Songs (${songs.length})',
+                style: Theme.of(context).textTheme.titleLarge,
               ),
             ),
             Expanded(
               child: ListView.builder(
-                padding: const EdgeInsets.fromLTRB(25.0, 0, 8.0, 8.0),
                 itemCount: songs.length,
                 itemBuilder: (context, index) {
                   final song = songs[index];
                   return ListTile(
+                    leading: const LeadingIcon(
+                      icon: Icon(Icons.music_note),
+                    ),
                     title: Text(song.title),
                     subtitle: Text(song.artist ?? ''),
                     onTap: () {
